Support for space greater then 4Gb. Updated space usage information on file create/delete.
Thanks to Wei-Ren Huang for the patch (added to contributors).



git-svn-id: http://boxfs.googlecode.com/svn/trunk@85 a81dc690-963a-11de-99ed-9d3e36ffeeb7

+++ boxapi.h
@@ -12,7 +12,7 @@ void 	api_free();
 int 	api_init(int* argc, char*** argv);
 
 int 	api_readdir(const char *, fuse_fill_dir_t, void * buf);
-void	api_getusage(long *, long * );
+void	api_getusage(long long *, long long * );
 
 int	api_open(const char *, const char *);
 int 	api_getattr(const char *path, struct stat *stbuf);
